+++ elebox-cli/src/manufacturer_cmd.rs
@@ -3,6 +3,10 @@ use elebox_core::Category;
 use elebox_core::Handler;
 use elebox_core::Manager;
 use elebox_core::Manufacturer;
+use std::io::stdin;
+use std::io::stdout;
+use std::io::Write;
+use std::ops::Not;
 
 #[derive(Debug, Args)]
 pub struct ManufacturerCommand {
@@ -19,7 +23,7 @@ enum ManufacturerSubCommand {
     Get(NameArgs),
 
     /// Remove a manufacturer from the database
-    Delete(NameArgs),
+    Delete(DeleteArgs),
 
     /// Update info of an existing manufacturer
     Update(UpdateArgs),
@@ -62,6 +66,15 @@ struct NameArgs {
 }
 
 #[derive(Debug, Args)]
+struct DeleteArgs {
+    name: String,
+
+    /// Skip confirm, delete directly
+    #[arg(short = 'Y', long = "yes")]
+    yes: bool,
+}
+
+#[derive(Debug, Args)]
 struct ExportArgs {
     #[arg(default_value = "elebox_export_mfrs.tsv")] // TODO filename
     path: String,
@@ -90,6 +103,22 @@ pub fn manufacturer_cmd(handler: elebox_core::ManufacturerHandler, cmd: &Manufac
             Err(err) => println!("Error: {err}"),
         },
         Some(ManufacturerSubCommand::Delete(args)) => {
+            // Confirm delete message
+            if !args.yes {
+                println!("Are you sure you want to delete '{}' ?", args.name);
+                print!("This action cannot be undone. [y/N]: ");
+
+                let mut input = String::new();
+                let _ = stdout().flush();
+                stdin().read_line(&mut input).expect("Failed to read input");
+
+                if input.trim_end().to_lowercase() != String::from("y") {
+                    println!("Deletion canceled");
+                    return;
+                }
+            }
+
+            println!("Deleting '{}'...", args.name);
             if let Err(err) = handler.delete(&args.name) {
                 println!("Error: {err}");
             };
