+++ elebox-cli/src/package_cmd.rs
@@ -5,6 +5,10 @@ use elebox_core::Manager;
 use elebox_core::Manufacturer;
 use elebox_core::Package;
 use elebox_core::PackageType;
+use std::io::stdin;
+use std::io::stdout;
+use std::io::Write;
+use std::ops::Not;
 
 #[derive(Debug, Args)]
 pub struct PackageCommand {
@@ -21,7 +25,7 @@ enum PackageSubCommand {
     Get(NameArgs),
 
     /// Remove a package from the database
-    Delete(NameArgs),
+    Delete(DeleteArgs),
 
     /// Update info of an existing package
     Update(UpdateArgs),
@@ -63,6 +67,15 @@ struct NameArgs {
 }
 
 #[derive(Debug, Args)]
+struct DeleteArgs {
+    name: String,
+
+    /// Skip confirm, delete directly
+    #[arg(short = 'Y', long = "yes")]
+    yes: bool,
+}
+
+#[derive(Debug, Args)]
 struct ExportArgs {
     #[arg(default_value = "elebox_export_packages.tsv")] // TODO filename
     path: String,
@@ -91,6 +104,22 @@ pub fn package_cmd(handler: elebox_core::PackageHandler, cmd: &PackageCommand) {
             Err(err) => println!("Error: {err}"),
         },
         Some(PackageSubCommand::Delete(args)) => {
+            // Confirm delete message
+            if !args.yes {
+                println!("Are you sure you want to delete '{}' ?", args.name);
+                print!("This action cannot be undone. [y/N]: ");
+
+                let mut input = String::new();
+                let _ = stdout().flush();
+                stdin().read_line(&mut input).expect("Failed to read input");
+
+                if input.trim_end().to_lowercase() != String::from("y") {
+                    println!("Deletion canceled");
+                    return;
+                }
+            }
+
+            println!("Deleting '{}'...", args.name);
             if let Err(err) = handler.delete(&args.name) {
                 println!("Error: {err}");
             };
